Update Nalu rules for multiple commits
It appears regression testing may not be applicable to the rank counts requested. Since NREL is able to generate all requested results with commit 1d3ee2, but Offerors may have generated most of their results with more recent versions, we will permit benchmark submissions with mixed code commits (a) at least for Nalu, and (b) in the event that a second code version (likely 1d3ee2) is needed to get around exceptions.